requirements.txt
Flask==2.3.3
Pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
openai==1.3.0
python-dotenv==1.0.0
werkzeug==2.3.7
